Refactored rewards to be given out only for luna removed claimpool, as it is no longer being used
